Add a still not working transcription using Google Speech API
There are some problems with TypeScript that I have no idea how to solve,
while the approach chosen for the async functions doesn't seem to be optimal.
Regardless of that, what's now transcriptor.ts worked well enough in JavaScript.

I don't know how to run it though, and made some changes on app.ts while trying to do so.

rxtranscriptor.ts is another approach to the async functions, using rxjs.
It's incomplete and also doesn't work at the moment, but I'll keep it.